+++ src/main/java/dooya/see/user/presentation/UserSignUpRequest.java
@@ -1,5 +1,6 @@
-package dooya.see.user.application;
+package dooya.see.user.presentation;
 
+import dooya.see.user.application.UserSignUpCommand;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -25,13 +26,4 @@ public record UserSignUpRequest(
         @NotBlank(message = "전화번호는 비어 있을 수 없습니다")
         String phoneNumber
 ) {
-        public UserSignUpCommand toCommand() {
-                return UserSignUpCommand.builder()
-                        .email(email)
-                        .name(name)
-                        .password(password)
-                        .birthDate(birthDate)
-                        .phoneNumber(phoneNumber)
-                        .build();
-        }
 }
