+++ src/main/java/dooya/see/user/presentation/UserSignUpResponse.java
@@ -1,0 +1,18 @@
+package dooya.see.user.presentation;
+
+import dooya.see.user.domain.Role;
+import dooya.see.user.domain.User;
+import lombok.Builder;
+
+import java.time.LocalDate;
+
+@Builder
+public record UserSignUpResponse(
+        Long id,
+        String email,
+        String name,
+        LocalDate birthDate,
+        String phoneNumber,
+        Role role
+) {
+}
